+++ app/src/main/java/com/uminari/practice/todoApp/views/TodoListFragment.kt
@@ -1,4 +1,8 @@
-package com.uminari.practice.views
+/*
+ * Copyright 2023 ryuji koyama
+ */
+
+package com.uminari.practice.todoApp.views
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -14,10 +18,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [TodoItemDetailFragment.newInstance] factory method to
+ * Use the [TodoListFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class TodoItemDetailFragment : Fragment() {
+class TodoListFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +39,7 @@ class TodoItemDetailFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.todo_item_fdetail_fragment, container, false)
+        return inflater.inflate(R.layout.todo_list_fragment, container, false)
     }
 
     companion object {
@@ -45,12 +49,12 @@ class TodoItemDetailFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment TodoItemDetailFragment.
+         * @return A new instance of fragment TodoListFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            TodoItemDetailFragment().apply {
+            TodoListFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
